`sct_register_to_template`: Move label functions into `labeling.py`
